aarch64: Fix aarch64 after moving to C23
This fixes a few aarch64 specific testcases after the move to default to GNU C23.
For the SME testcases, the GNU C23 cases as `()` changing to mean `(void)` instead
of a non-prototype declaration; the non-prototype declaration merging was confusing
some of the time so the updated way is the expected way even for that.
For pic-*.c `-Wno-old-style-definition` was added not to warn about old style definitions.
For pr113573.c, I added `-std=gnu17` since I was not sure if `(...)` with C23 would invoke
the same issue.

tested for aarch64-linux-gnu.

	PR testsuite/117680
gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/pic-constantpool1.c: Add -Wno-old-style-definition.
	* gcc.target/aarch64/pic-symrefplus.c: Likewise.
	* gcc.target/aarch64/pr113573.c: Add `-std=gnu17`
	* gcc.target/aarch64/sme/streaming_mode_1.c: Correct testcase.
	* gcc.target/aarch64/sme/za_state_1.c: Likewise.
	* gcc.target/aarch64/sme/za_state_2.c: Likewise.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>