RISC-V: Refine the vector stride load/store testcases
The rtl expand dump for IFN check of stride load/store testcase is
different between O2 and O3.  It it reasonable to leverage target
no-opts/any-opts to filte out, instead of the xfail.

The below test suites are passed for this patch.
* The rv64gcv fully regression test.

It is test only patch and obvious up to a point, will commit it
directly if no comments in next 48H.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-f16.c: Take
	the target any-ops instead of xfail.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-f32.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-f64.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-i16.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-i32.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-i64.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-i8.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-u16.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-u32.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-u64.c: Ditto.
	* gcc.target/riscv/rvv/autovec/strided/strided_ld_st-1-u8.c: Ditto.

Signed-off-by: Pan Li <pan2.li@intel.com>