ada: Elide the copy for bit-packed aggregates in object declarations
The in-place expansion has been historically disabled for them, but there
does not seem to be any good reason left for this.  However, this requires
a small trick in order for the expanded code not to be flagged as using the
object uninitialized by the code generator.

gcc/ada/ChangeLog:

	* exp_aggr.adb (Convert_Aggr_In_Object_Decl): Clear the component
	referenced on the right-hand side of the first assignment generated
	for a bit-packed array, if any.
	(Expand_Array_Aggregate): Do not exclude aggregates of bit-packed
	array types in object declarations from in-place expansion.
	* sem_eval.adb (Eval_Indexed_Component): Do not attempt a constant
	evaluation for a bit-packed array type.