ada: Defend against risk of infinite loop
A recently fixed bug caused an infinite loop when assertions were not
checked. With assertions checked, the symptom was just an internal
error caused by an assertion failure. This patch makes it so that if
another bug ever causes the same condition to fail, there will never be
an infinite loop with any assertion policy.

gcc/ada/ChangeLog:

	* sem_ch3.adb (Access_Subprogram_Declaration): Replace assertion with
	more defensive code.