c++: modules and class attributes
std/time/traits/is_clock.cc was getting a warning about applying the
deprecated attribute to a variant of auto_ptr, which was wrong because it's
on the primary type.  This turned out to be because we were ignoring the
attributes on the definition of auto_ptr because the forward declaration in
unique_ptr.h has no attributes.  We need to merge attributes as usual in a
redeclaration.

gcc/cp/ChangeLog:

	* module.cc (trees_in::decl_value): Merge attributes.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/attrib-1_a.C: New test.
	* g++.dg/modules/attrib-1_b.C: New test.