build: Discard obsolete references to $(GCC_PARTS)
The $(GCC_PARTS) variable was deleted with the Makefile rework in commit
fa9585134f6f ("libgcc move to the top level")[1] back in 2007, and yet
the Ada and Modula 2 frontends added references to this variable later
on, with commit e972fd5281b7 ("[Ada] clean ups in Makefiles")[2] back in
2011 and commit 1eee94d35177 ("Merge modula-2 front end onto gcc.") back
in 2022 respectively.

I guess it's because the frontends lived too long externally.  Discard
the references then, they serve no purpose nowadays.

References:

[1] <https://inbox.sourceware.org/gcc-patches/20061219191730.GA32355@nevyn.them.org/>

[2] <https://inbox.sourceware.org/gcc-patches/20110906151510.GA5213@adacore.com/>

	gcc/ada/
	* gcc-interface/Make-lang.in (gnattools): Remove $(GCC_PARTS).

	gcc/m2/
	* Make-lang.in (m2 modula-2 modula2): Remove $(GCC_PARTS).