match: Improve handling of double convert [PR117776]
For a double conversion, we will simplify it into a conversion
with an and if the outer type and inside precision matches and
the intra precision is smaller and unsigned. We should be able
to extend this to where the outer precision is larger too.
This is a good canonicalization too.

Bootstrapped and tested on x86_64-linux-gnu.

	PR tree-optimization/117776
gcc/ChangeLog:

	* match.pd (nested int casts): Allow for the case
	where the final prec is greater than the original
	prec.

gcc/testsuite/ChangeLog:

	* g++.dg/vect/pr117776.cc: New test.
	* gcc.dg/tree-ssa/cast-3.c: New test.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>