c++: modules, generic lambda, constexpr if
In std/ranges/concat/1.cc we end up instantiating
concat_view::iterator::operator-, which has nested generic lambdas, where
the innermost is all constexpr if.  tsubst_lambda_expr propagates
the returns_* flags for generic lambdas since we might not substitute into
the whole function, as in this case with constexpr if.  But the module
wasn't preserving that flag, and so the importer gave a bogus "no return
statement" diagnostic.

gcc/cp/ChangeLog:

	* module.cc (trees_out::write_function_def): Write returns* flags.
	(struct post_process_data): Add returns_* flags.
	(trees_in::read_function_def): Set them.
	(module_state::read_cluster): Use them.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/constexpr-if-1_a.C: New test.
	* g++.dg/modules/constexpr-if-1_b.C: New test.